// src/type/ViewLib.d.ts
export interface ViewLib {
  name: string,
  groupName: string,
  deviceType: string,
  latitude: number,
  longitude: number,
  placeCode: string,
  orgCode: string,
  model: string,
  viidServerId: string,
  inProtocol?: string,
  ipAddr: string,
  groupId: number,
  place?: string,
  description?: string,
  status: number,
  certId: number,
  certName?: string,
  ipv6Addr?: string,
  port: number,
  viidServerIp: string,
  viidServerPort: number,
  apeType: string,
}